banco de dados de resultado adicionado, células da recycler view customizadas, criada janela que mostra todas os registros salvos de todos os tipos e arrumada a descrição do info IMC

+++ app/src/main/java/com/nnt/calculadoraimc_turmac/FrequenciaCardiacaMaximaResultActivity.kt
@@ -14,7 +14,7 @@ class FrequenciaCardiacaMaximaResultActivity : AppCompatActivity() {
         binding = ActivityFrequenciaCardiacaMaximaResultBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        binding.textViewResultadoTMB.text=intent.getIntExtra("fcMaxima",0).toString()
+        binding.textViewResultadoTMB.text=intent.getDoubleExtra("fcMaxima",0.0).toInt().toString()
 
         binding.imageButtonVoltar.setOnClickListener {
             startActivity(Intent(this,FrequenciaCardiacaMaximaActivity::class.java))
